refactor(oci): decompose the fetch god-function into a readable pipeline
pkg/source/oci worked but read "all over the place": fetch() was a
184-line god function inlining ref parse, credential/TLS/transport/auth
wiring, semver + digest resolution, resolve-cache, slot acquisition, the
cache-hit gauntlet, oras.Copy, cosign verify, layer extract, ignore
filter, and marker writes. Cache bookkeeping was tangled into the pull
pipeline, the "OCIRepository ns/name" error prefix was hand-formatted
~15×, the two marker writers duplicated the same read-modify-write, and
layer-selector logic was smeared across three files.

Pure internal refactor — behavior-preserving, byte-equivalent, no public
API change (only Fetcher/Fetch stay exported):

- client.go (new): registry-client construction — newRepoClient (extracted
  from fetch's inline setup) + TLS/registry-config/credentials (from the
  now-deleted auth.go).
- cache.go (new): cache bookkeeping moved out of fetch.go — ociCacheKey,
  ociResolveCacheKey, lookupResolveCache, writeResolveCache, checkCacheHit,
  ociArtifact, with the resolve-prefix / opts-separator as named constants.
- fetch.go: fetch() is now a ~75-line pipeline (validate → newRepoClient →
  resolveRef → resolve digest → slot → cache-hit → copyArtifact →
  publishArtifact); copyArtifact and publishArtifact are the extracted
  stages. The slot transaction (Release/Refresh/Commit ordering, the
  verified-gated marker write) is preserved exactly.
- marker.go: writeCachedDigest/writeVerifyMarker collapse onto one
  updateSlotMeta(slot, mutate) read-modify-write.
- layer.go: layerMediaType + the helm media-type constants now live with
  the other layer-selection code (moved from resolve.go).
- fetcher.go: ociID(repo) helper unifies the error prefix across the
  package; resolve.go gains latestTag + resolveOCIRefDigest.

Verification: gofmt/vet/lint clean, full suite + -race pass (the existing
httptest fake-registry suite exercises every cache-key / semver / layer /
verify-marker / slot-transaction invariant), and the 24-repo cross-cluster
byte-equivalence sweep is 24/24 identical to main. Net -248 lines.
